% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_eh.R
\name{calc_eh}
\alias{calc_eh}
\title{Calculate expected heterozygosity}
\usage{
calc_eh(genotypes)
}
\arguments{
\item{genotypes}{a matrix of haploid markers as rows with alleles encoded as
0 and 2}
}
\value{
a vector of expected heterozygosities of the markers
}
\description{
Calculates expected heterozygosity of markers given a matrix where rows are
markers, and markers are haploid, encoded as 0 and 2
}
