% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_ld_stats.R
\name{calc_ld_stats}
\alias{calc_ld_stats}
\title{Calculate composite ld between markers}
\usage{
calc_ld_stats(gds, snp_data)
}
\arguments{
\item{gds}{a gds object}

\item{snp_data}{the parsed snp data of the gds object}
}
\value{
a list contianing the dist object and neighbouring lds for each
chromosome ordered by genome
}
\description{
Calculates composite ld for the markers on each chromosome in a gds object,
returning a dist object and the LD between neighbouring markers
}
