% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_lng.R
\name{calc_lng}
\alias{calc_lng}
\title{Calculate LNG}
\usage{
calc_lng(snp_data)
}
\arguments{
\item{snp_data}{the parsed snp data of the gds object}
}
\value{
a list contianing the above data organised by genome
}
\description{
Calculates the length of each chromsome in each genome, the number of
markers and the gaps between neighbouring markers
}
