% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_maf_mr.R
\name{calc_maf_mr}
\alias{calc_maf_mr}
\title{Calculate the MAF and MR}
\usage{
calc_maf_mr(wheat_data)
}
\arguments{
\item{wheat_data}{the parsed gds object}
}
\value{
a list contianing the above data organised by genome
}
\description{
Calculates the minor allele frequency (MAF) and missing rate (MR) of each
marker
}
