% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_coverage_by_chrom.R
\name{coverage_by_chrom}
\alias{coverage_by_chrom}
\title{Calculate mean allele richness at all sampling levels}
\usage{
coverage_by_chrom(snp_chrom, snp_pos)
}
\arguments{
\item{snp_chrom}{a factor vector inidicating the chrom of each marker}

\item{snp_pos}{a vector with the position of each snp on its chrom}
}
\value{
a table of expected allele richness for each marker at each 
subsampling level with markers in rows and sampling levels in columns
}
\description{
Calculates the mean allele richness across all markers for a sample at all
sampling levels. Based on the formula presented in 
https://www.genetics.org/content/157/1/389
}
