% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_rects.R
\name{draw_rects}
\alias{draw_rects}
\title{Craws rectangles for circliz graph}
\usage{
draw_rects(pop_code, colour_subset, label_order, border)
}
\arguments{
\item{pop_code}{a factor of the classifications of the individuals}

\item{colour_subset}{the colours used for the factor levels}

\item{border}{the colour for the border of the rectangles}

\item{the}{order of the factors in the row}
}
\value{
None
}
\description{
Function for drawing rectangles in each track of a cirlcize graph
}
