% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_genes.R
\name{load_genes}
\alias{load_genes}
\title{Loads gene info}
\usage{
load_genes(csv, base = 0)
}
\arguments{
\item{csv}{the csv file containg the alignments}

\item{base}{used further on as the y axis position of the genes during
plotting}
}
\value{
a tibble contianign the parsed info
}
\description{
Loads the info on the selected gene alignments
}
