% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_max_lengths.R
\name{max_lengths}
\alias{max_lengths}
\title{Calc genome and homoeologous set lengths from chromosomes}
\usage{
max_lengths(chrom_lengths)
}
\arguments{
\item{chrom_lengths}{a list of lengths by chrom in numeric order}
}
\value{
a list containing the lengths of the different genome and 
homoeolog sets
}
\description{
Calc genome and homoeologous set lengths from chromosomes
}
