% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/round_df.R
\name{round_df}
\alias{round_df}
\title{From: https://jeromyanglim.tumblr.com/post/50228877196/round-numbers-in-data-frame-that-contains-non}
\usage{
round_df(x, digits)
}
\description{
From: https://jeromyanglim.tumblr.com/post/50228877196/round-numbers-in-data-frame-that-contains-non
}
