% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snpgds_parse.R
\name{snpgds_parse}
\alias{snpgds_parse}
\title{Parse a gds file}
\usage{
snpgds_parse(gds_file)
}
\arguments{
\item{gds_file}{the path to the target gds file}
}
\value{
a list of lists containing the data of the gds object
}
\description{
Loads the data from a gds object into a form amenable to manipulation
using R code
}
