% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snpgds_sample_subset.R
\name{snpgds_sample_subset}
\alias{snpgds_sample_subset}
\title{Create a gds file that cotnains a subset of the sample from another gds file}
\usage{
snpgds_sample_subset(wheat_data, gds_file, sample_index)
}
\arguments{
\item{wheat_data}{the parsed gds data}

\item{gds_file}{the path and name of the new gds file}

\item{sample_index}{the indices of the samples to be removed}
}
\value{
None
}
\description{
Takes the data parsed from a gds file using snpgds_parse and creates a new
gds file without the samples given by sample_index
}
