% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snpgds_snp_subset.R
\name{snpgds_snp_subset}
\alias{snpgds_snp_subset}
\title{Create a gds file that contains a subset of the snps of another gds file}
\usage{
snpgds_snp_subset(wheat_data, gds_file, snp_index)
}
\arguments{
\item{wheat_data}{the parsed gds data}

\item{gds_file}{the path and name of the new gds file}

\item{snp_index}{the indices of the snps to be removed}
}
\value{
None
}
\description{
Takes the data parsed from a gds file using snpgds_parse and creates a new
gds file without the snps given by snpindex
}
