% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_span_by_chrom.R
\name{span_by_chrom}
\alias{span_by_chrom}
\title{Calculate the length of each chrom, optionally subtract the min position from
it}
\usage{
span_by_chrom(snp_chrom, snp_pos, diff = FALSE)
}
\arguments{
\item{snp_chrom}{a factor vector inidicating the chrom of each marker}

\item{snp_pos}{a vector with the position of each snp on its chrom}
}
\value{
a vector of span distances
}
\description{
Calculate the length of each chrom, optionally subtract the min position from
it
}
