% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_allele_richness.R
\name{allele_richness}
\alias{allele_richness}
\title{Calculate mean allele richness at all sampling levels}
\usage{
allele_richness(pop, allele_coding = 1:2, num_cores = detectCores())
}
\arguments{
\item{pop}{a data frame with individuals in columns and markers in rows,
there must be atleast two individuals, missing data is not allowed}

\item{allele_coding}{the coding used for indicating the different alleles}

\item{num_cores}{the number of cores to use with mclapply(), must be 1 on
windows, default is detectCores()}
}
\value{
a table of expected allele richness for each marker at each 
subsampling level with markers in rows and sampling levels in columns
}
\description{
Calculates the mean allele richness across all markers for a sample at all
sampling levels. Based on the formula presented in 
https://www.genetics.org/content/157/1/389
}
