% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_allele_richness.R
\name{allele_richness}
\alias{allele_richness}
\title{Calculate mean allele richness at all sampling levels}
\usage{
allele_richness(pop, coding = 1:2, num_cores = 1)
}
\arguments{
\item{pop}{a data frame with individuals in columns and markers in rows}

\item{coding}{the coding used for indicating the different alleles}

\item{num_cores}{the number of cores to use, must be 1 on windows, can use 
detectCores of the parallel package on linux}
}
\value{
a table of expected allele richness for each marker at each 
subsampling level with markers in columns and sampling levels in rows
}
\description{
Calculates the mean allele richness across all markers for a sample at all
sampling levels. Missing data is not allowed. Based on the formula presented
in https://www.genetics.org/content/157/1/389
}
