% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_plot_map_stats.R
\name{calc_plot_map_stats}
\alias{calc_plot_map_stats}
\title{Calculate and plot map stats}
\usage{
calc_plot_map_stats(gds, plot_title, y_lim, out_name)
}
\arguments{
\item{gds}{the file path to the gds object}

\item{plot_title}{the title of the combined plots}

\item{y_lim}{the limit for the y axis of the plots}

\item{out_name}{the base name to use for the csv and png outputs}
}
\value{
Outputs plots and tables of the data
}
\description{
Calcluates various map statistics creating plots and tables
}
