% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_gaps_nbs_ld.R
\name{plot_gaps_nbs_ld}
\alias{plot_gaps_nbs_ld}
\title{Output figures based on map stats}
\usage{
plot_gaps_nbs_ld(lng, genome_ld, gds, plot_title, y_lim, out_name)
}
\arguments{
\item{lng}{data supplied by the calc_lng function}

\item{genome_ld}{data supplied by the calc_cld_stats function}

\item{gds}{the gds object}

\item{plot_title}{the title of the plot}

\item{y_lim}{the y-axis limit of the plot}

\item{out_name}{the name for the output file}
}
\value{
None
}
\description{
Outputs a figure of gap distances and neighbouring ld by genome and overall
}
