% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_span_by_chrom.R
\name{span_by_chrom}
\alias{span_by_chrom}
\title{Calculate the difference between the max and min positions on a chromosome}
\usage{
span_by_chrom(snp_pos, snp_chrom)
}
\arguments{
\item{snp_pos}{a vector with the position of each snp on its chrom}

\item{snp_chrom}{a factor vector inidicating the chrom of each marker}
}
\value{
a vector of span distances
}
\description{
Calculate the difference between the max and min positions on a chromosome
}
